Enhancement: More Flexible Docs Builder
The old build-docs-by-type never really had the flexibility that we had
hoped and there was a need to handle custom API docs builder. This PR
makes the docs builder work more in line with the other commands where a
reasonable default is set but a user can add a list of commands in the
pyproject.toml file.