Enhancement: Allow KeyboardInterrupt
When running some commands, the command could actually include an event
loop that waits for a user to exit (e.g., `sphinx-autobuild`). Right
now, when the user exits out via Ctrl+C, a nasty traceback is displayed.
This change will catch that error so that it is not shown to the user.