Bugfix: Always Install Dependencies
Hit a condition that I have encountered on other projects where the
cache check found an invalid cache. As a result, changing to always run
the `poetry install` step. Worst case, it takes just a couple of seconds
to say "all is good". Best case, it updates the environment with the bad
cache.