Enhancement: `tomlkit` to `tomli`
With the acceptance of PEP-680
(https://www.python.org/dev/peps/pep-0680/), it looks like `tomli` is
going to be the basis for `tomllib` in the standard library. The library
is a very lightweight library with strong testing and no dependencies,
fitting our needs well.

In addition, bump the base Python version to 3.7 since 3.6 reached
End-of-Life in December 2021
(https://www.python.org/dev/peps/pep-0494/#lifespan).